cors enabled, fix some error NOT checked

+++ src/main/java/valerio/BingeBookBE/entity/Film.java
@@ -6,7 +6,6 @@ import lombok.Getter;
 import lombok.NoArgsConstructor;
 import lombok.Setter;
 
-import java.math.BigInteger;
 import java.util.HashSet;
 import java.util.Set;
 
@@ -19,9 +18,9 @@ import com.fasterxml.jackson.annotation.JsonIgnore;
 public class Film {
     @Id
     @GeneratedValue
-//     @Getter(AccessLevel.NONE)
+    // @Getter(AccessLevel.NONE)
     @Setter(AccessLevel.NONE)
-    private BigInteger id;
+    private Long id;
 
     private String title;
 
@@ -29,15 +28,11 @@ public class Film {
     private String posterUrl;
 
     @ManyToMany
-    @JoinTable(name = "film_genre",
-            joinColumns = @JoinColumn(name = "film_id"),
-            inverseJoinColumns = @JoinColumn(name = "genre_id"))
+    @JoinTable(name = "film_genre", joinColumns = @JoinColumn(name = "film_id"), inverseJoinColumns = @JoinColumn(name = "genre_id"))
     private Set<Genre> genres = new HashSet<>();
 
     @ManyToMany
-    @JoinTable(name = "film_tag",
-            joinColumns = @JoinColumn(name = "film_id"),
-            inverseJoinColumns = @JoinColumn(name = "tag_id"))
+    @JoinTable(name = "film_tag", joinColumns = @JoinColumn(name = "film_id"), inverseJoinColumns = @JoinColumn(name = "tag_id"))
     private Set<Tag> tags = new HashSet<>();
 
     @JsonIgnore
