+++ src/main/java/valerio/BingeBookBE/entity/Genre.java
@@ -10,7 +10,6 @@ import lombok.Getter;
 import lombok.NoArgsConstructor;
 import lombok.Setter;
 
-import java.math.BigInteger;
 import java.util.HashSet;
 import java.util.Set;
 
@@ -21,9 +20,9 @@ import java.util.Set;
 public class Genre {
     @Id
     @GeneratedValue
-//    @Getter(AccessLevel.NONE)
+    // @Getter(AccessLevel.NONE)
     @Setter(AccessLevel.NONE)
-    private BigInteger id;
+    private Long id;
     private String name;
 
     @JsonIgnore
